refactoring for fx and biz
