--- 07rpc/rpc01/rpcfx-client/src/main/resources/application.yml
@@ -1,2 +1,0 @@
-#server:
-#  port: 8080
