--- 07rpc/rpc01/rpcfx-client/src/test/java/io/kimmking/rpcfx/client/Springboot01ApplicationTests.java
@@ -1,13 +1,0 @@
-//package io.kimmking.springboot01;
-//
-//import org.junit.Test;
-//import org.springframework.boot.test.context.SpringBootTest;
-//
-//@SpringBootTest
-//class Springboot01ApplicationTests {
-//
-//	@Test
-//	void contextLoads() {
-//	}
-//
-//}
