+++ 07rpc/rpc01/rpcfx-core/src/main/java/io/kimmking/rpcfx/api/RpcfxResolver.java
@@ -1,0 +1,7 @@
+package io.kimmking.rpcfx.api;
+
+public interface RpcfxResolver {
+
+    Object resolve(String serviceClass);
+
+}
