+++ 07rpc/rpc01/rpcfx-demo-provider/src/main/java/io/kimmking/rpcfx/demo/provider/DemoResolver.java
@@ -1,0 +1,20 @@
+package io.kimmking.rpcfx.demo.provider;
+
+import io.kimmking.rpcfx.api.RpcfxResolver;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
+
+public class DemoResolver implements RpcfxResolver, ApplicationContextAware {
+
+    private ApplicationContext applicationContext;
+
+    @Override
+    public void setApplicationContext(ApplicationContext applicationContext) {
+        this.applicationContext = applicationContext;
+    }
+
+    @Override
+    public Object resolve(String serviceClass) {
+        return this.applicationContext.getBean(serviceClass);
+    }
+}
