+++ 07rpc/rpc01/rpcfx-demo-provider/src/main/java/io/kimmking/rpcfx/demo/provider/OrderServiceImpl.java
@@ -1,4 +1,4 @@
-package io.kimmking.rpcfx.server;
+package io.kimmking.rpcfx.demo.provider;
 
 import io.kimmking.rpcfx.demo.api.Order;
 import io.kimmking.rpcfx.demo.api.OrderService;
