+++ 07rpc/rpc01/rpcfx-demo-provider/src/main/java/io/kimmking/rpcfx/demo/provider/UserServiceImpl.java
@@ -1,8 +1,7 @@
-package io.kimmking.rpcfx.server;
+package io.kimmking.rpcfx.demo.provider;
 
 import io.kimmking.rpcfx.demo.api.User;
 import io.kimmking.rpcfx.demo.api.UserService;
-import io.kimmking.rpcfx.demo.api.User;
 
 public class UserServiceImpl implements UserService {
